feat(copy): add support for file copying operations in database (#75)
* feat(copy): add support for file copying operations in database

Added files and functions necessary for implementing the COPY command in the database, allowing for both copying to and from files. This includes the creation of the CopyFromFile and CopyToFile operators and executors, modifications to the binder and executor mod files to support the new COPY command structure, and the creation of several associated test files. Additionally, made necessary changes to the CreateTable executor to yield tuples as the execution result, and introduced the 'csv' dependency in Cargo.toml for handling CSV file reading and writing. This feature allows for efficient data importation and exportation between database tables and external files.

* remove the he unwrap

* remove duplicate module import in operator

* cargo fmt

* style: added PrimaryKeyNotFound error

* fix: handle error process

---------

Co-authored-by: Kould <2435992353@qq.com>